Move code to lib.rs and create get_scores
At the moment get_scores takes a Vec with all the cities (departing,
arriving and in between) and naively checks if tickets are
valid. Valid tickets are the ones that both destinations are included
as values in the vector. Still I need to implement the score from
using trains, and the +10 for longest route