Add bool in dijkstra and search for longest returned route.
I've added the bool argument so that I can differentiate when to stop
at the destination or not. Another way to do that without the bool
would be to use the same function with different arguments? I need to
check how this can be done in Rust.

Also I check to find the city with the greatest distance from
Edinburgh, which turns out to be Erzurum. So my plan is to check all
the available routes and compute their value, to find another possible
largest theoretical score.

Talking about score, I need to change the score computing function to
take into account also the value of the trains used.