Factor out the max_key function and find the longest route
I created the generic max_key function to use it later multiple times
and simplify the code. Also I added some logic to find the longest
route discovered from Dijkstra function. Still, I need to find the
longest route that also includes City::Athina, so that I can maximize
the score of the route. Still Dijkstra is only an experiment, I need
to find the longest possible route that connects Endinburgh with
Athens and gives the greatest score with the tickets covered. This
will be my next step.