Use "fully" qualified names
I think appending the module name in front of the function makes it
easier for someone that wants peruse the code to find the source of
the function. I think routing::dijkstra(..) is better than
dijkstra(..) as you can see with a glance in which module you are
going to find the dijkstra function in the first case.