Count trains used also in scoring
Till now, for scoring I only checked if the two points of a ticket
existed in the given route, and if yes, the score was increased
accordingly. But I think that the max score may go over 255, so I
guard for this. I'm not sure if the maximum score, now that I think of
it, may go over 255, but just to be on the safe side, I converted the
score to 16 bit unsigned, instead of 8 bit unsigned. If this is not
the case, I may bring it back to u8.