First attempt to traverse the network
This is my first attempt. I really tried and succeeded to make the
compiler happy, but unfortunately, the result was not what I
expected. I thought of keeping in a set in a hashmap the cities
visited by a specific city, but the problem was that loops were
created that were not what I wanted. Nevertheless this was a good
first try. Maybe I'll try recursion next time.

Maybe the solution is to check if a child node is visited from the
current node, and if the node exists already in the route. So even if,
for example, I haven't visited Amsterdam from Berlin, if Amsterdam is
already in my route, I don't need to go there again, so I'll have to
choose another child of Berlin. Yes, that's probably a solution to my
problem. Till next time!