Full network traversal is done?
I fixed the problems I had. That is, I have added the condition that
the next neighbor to visit is not in the visited hashmap and it's not
in the current_route either, so that I avoid loops to same cities.

Also I found out why progressively the routes were smaller and smaller
and ending after about 30 routes. When I reach a dead_end, I retract
to previous node with current_route.pop(), but also I need to clear
the visited node of the dead_end, so that if I visit it from another
node, it's neighbors are again available.

The surprise initially was that u16 was not enough for the number of
routes so I upgraded the route_id to u32 and now the number of routes
found is about 2 million.