Cut on routes stores for scoring
This is an easy change, after adding debug traces, I noticed that a
lot of small routes were added to the routes vector, which increased
the number of small routes to compute scores. Now I check that the
route completes a big ticket and that there are no more that 2 trains
remaining unused.